Factoring out of endpoints in individual files
